requirement.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.5
django-admin-tools==0.9.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-environ-plus==0.9.0.1
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.12.0
pygraphviz==1.9
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
wcwidth==0.2.5
